requirements.txt
flask
python-dotenv
transformers
torch
torchvision
numpy
matplotlib